Move AsyncActor module to Stackdriver::Core. Refactor debugger actor classes.
