+++ google-cloud-debugger/test/google/cloud/debugger/middleware_test.rb
@@ -30,6 +30,8 @@ describe Google::Cloud::Debugger::Middleware, :mock_debugger do
   after {
     Google::Cloud::Debugger.configure.instance_variable_get(:@configs).clear
     Google::Cloud.configure.delete :use_debugger
+
+    debugger.stop
   }
 
   describe "#call" do
