+++ stackdriver-core/test/helper.rb
@@ -19,3 +19,16 @@ require "minitest/rg"
 require "ostruct"
 require "json"
 require "base64"
+
+##
+# Helper method to loop until block yields true or timeout.
+def wait_until_true timeout = 5
+  begin_t = Time.now
+
+  until yield
+    return :timeout if Time.now - begin_t > timeout
+    sleep 0.1
+  end
+
+  :completed
+end
