+++ stackdriver-core/test/stackdriver/core/async_actor_test.rb
@@ -14,13 +14,16 @@
 
 
 require "helper"
+require "stackdriver/core/async_actor"
 
-describe Google::Cloud::Debugger::AsyncActor do
+describe Stackdriver::Core::AsyncActor do
   class AsyncActorTest
-    include Google::Cloud::Debugger::AsyncActor
+    include Stackdriver::Core::AsyncActor
 
     def initialize
       super()
+
+      set_cleanup_options timeout: 0.1
     end
 
     def run_backgrounder
@@ -184,8 +187,10 @@ describe Google::Cloud::Debugger::AsyncActor do
     it "waits for the async job to stop" do
       actor.async_start
 
+      actor.send :set_cleanup_options, force: false
+
       actor.async_stopped?.must_equal false
-      stop = actor.async_stop! 0.1
+      stop = actor.async_stop!
       stop.must_equal :waited
       actor.async_stopped?.must_equal true
     end
@@ -204,7 +209,18 @@ describe Google::Cloud::Debugger::AsyncActor do
       wait_result.must_equal :completed
 
       actor.async_stopped?.must_equal false
-      stop = actor.async_stop! 0.1, force: true
+      stop = actor.async_stop!
+      stop.must_equal :forced
+      actor.async_stopped?.must_equal true
+    end
+
+    it "doesn't wait if timeout is 0" do
+      actor.async_start
+
+      actor.send :set_cleanup_options, timeout: 0
+
+      actor.async_stopped?.must_equal false
+      stop = actor.async_stop!
       stop.must_equal :forced
       actor.async_stopped?.must_equal true
     end
@@ -212,7 +228,7 @@ describe Google::Cloud::Debugger::AsyncActor do
 
   describe ".register_for_cleanup" do
     it "adds actor to cleanup_list" do
-      klass = Google::Cloud::Debugger::AsyncActor
+      klass = Stackdriver::Core::AsyncActor
       actor.async_start
       klass.instance_variable_get("@cleanup_list").must_include actor
       actor.async_stop
@@ -222,12 +238,10 @@ describe Google::Cloud::Debugger::AsyncActor do
   describe ".run_cleanup" do
     it "calls async_stop! on actors" do
       mock = Minitest::Mock.new
-      mock.expect :async_stop!, true, [
-                  Google::Cloud::Debugger::AsyncActor::CLEANUP_TIMEOUT,
-                  {force: true}]
+      mock.expect :async_stop!, true, []
 
-      Google::Cloud::Debugger::AsyncActor.register_for_cleanup mock
-      Google::Cloud::Debugger::AsyncActor.send :run_cleanup
+      Stackdriver::Core::AsyncActor.register_for_cleanup mock
+      Stackdriver::Core::AsyncActor.send :run_cleanup
 
       mock.verify
     end
